fix: organize layouts
move card and stack from organism to layouts
